refactor: move stateful logic to custom hook returning handler function as prop from the container compoent for better testability
relates #393
